WIP Porting: Relocated base_sequence in sequences folder
