+++ ecpy_pulses/pulses/sequences/views/conditional_view.enaml
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
-# =============================================================================
-# module : conditional_view.enaml
-# author : Matthieu Dartiailh
-# license : MIT license
-# =============================================================================
+# -----------------------------------------------------------------------------
+# Copyright 2015-2016 by EcpyPulses Authors, see AUTHORS for more details.
+#
+# Distributed under the terms of the BSD license.
+#
+# The full license is in the file LICENCE, distributed with this software.
+# -----------------------------------------------------------------------------
 """
 """
 from enaml.layout.api import hbox, align, spacer, vbox
